enigma_dyn: add zapStream.m3u support (by ims)
This adds functionality to have enigma zap to a certain
channel, and start streaming it.
It does not use the 'zapstream' application, as the name
might lead to suspect; it works the other way around.
Instead of the stream following when enigma zaps, the
stream dictates enigma where to zap to.